feat(auth): Separating context from rest
