+++ module/auth-rest-jwt/src/principal-encoder.ts
@@ -1,5 +1,5 @@
-import { Principal } from '@travetto/auth';
-import { AuthService, PrincipalEncoder } from '@travetto/auth-rest';
+import { AuthContextService, Principal } from '@travetto/auth';
+import { PrincipalEncoder } from '@travetto/auth-rest';
 import { AppError, Runtime, TimeSpan, TimeUtil } from '@travetto/runtime';
 import { Config } from '@travetto/config';
 import { Inject, Injectable } from '@travetto/di';
@@ -49,7 +49,7 @@ export class JWTPrincipalEncoder implements PrincipalEncoder {
   config: RestJWTConfig;
 
   @Inject()
-  auth: AuthService;
+  auth: AuthContextService;
 
   toJwtPayload(p: Principal): Payload {
     const exp = TimeUtil.asSeconds(p.expiresAt!);
