+++ module/auth-rest-jwt/support/test/server.ts
@@ -4,8 +4,8 @@ import timers from 'node:timers/promises';
 import { Controller, Get, Post, Redirect, Request } from '@travetto/rest';
 import { Suite, Test } from '@travetto/test';
 import { DependencyRegistry, Inject, InjectableFactory } from '@travetto/di';
-import { AuthenticationError, Authenticator } from '@travetto/auth';
-import { AuthService, Authenticate, Authenticated } from '@travetto/auth-rest';
+import { AuthContextService, AuthenticationError, Authenticator } from '@travetto/auth';
+import { Authenticate, Authenticated, LoginService } from '@travetto/auth-rest';
 import { JWTUtil } from '@travetto/jwt';
 
 import { BaseRestSuite } from '@travetto/rest/support/test/base';
@@ -37,7 +37,10 @@ class Config {
 class TestAuthController {
 
   @Inject()
-  svc: AuthService;
+  svc: AuthContextService;
+
+  @Inject()
+  login: LoginService;
 
   @Post('/login')
   @Authenticate(TestAuthSymbol)
@@ -59,7 +62,7 @@ class TestAuthController {
   @Get('/logout')
   @Authenticated()
   async logout(req: Request) {
-    await this.svc.logout(req);
+    await this.login.logout(req);
     return new Redirect('/auth/self', 301);
   }
 }
