+++ module/auth-rest/src/internal/types.ts
@@ -2,9 +2,6 @@ import { PrincipalTarget } from '@travetto/auth/src/internal/types';
 import { ContextProvider, ParamExtractor } from '@travetto/rest';
 
 export const LoginContextSymbol = Symbol.for('@travetto/auth-rest:login');
-export const AuthTokenSymbol = Symbol.for('@travetto/auth-rest:token');
-
-export type AuthToken = { token: string, type: string };
 
 /**
  * @augments `@travetto/rest:Context`
