+++ module/auth/__index__.ts
@@ -1,4 +1,5 @@
 export * from './src/util';
+export * from './src/service';
 export * from './src/types/authenticator';
 export * from './src/types/authorizer';
 export * from './src/types/principal';
