+++ module/auth/src/internal/types.ts
@@ -10,3 +10,8 @@ export class PrincipalTarget {
 export class AuthorizerTarget { }
 export class AuthenticatorTarget { }
 export class PrincipalSerializerTarget { }
+
+export const AuthTokenSymbol = Symbol.for('@travetto/auth:token');
+export const PrincipalSymbol = Symbol.for('@travetto/auth:principal');
+
+export type AuthToken = { token: string, type: string };
